Replace separate repository tests with integration test
